Add a restructuration on code.
