+++ src/js/mockData/tasks.js
@@ -1,0 +1,23 @@
+const firstTasks = [
+    {
+      index: 1,
+      description: 'Go to the store',
+      completed: false,
+    },
+    {
+      index: 2,
+      description: 'Finish my screencast',
+      completed: false,
+    },
+    {
+      index: 3,
+      description: 'Clean the house',
+      completed: true,
+    },
+    {
+      index: 4,
+      description: 'Mow the lawn',
+      completed: false,
+    },
+];
+module.export = firstTasks;
